Change to coming soon site
Because of a move to Ohio, I am not able to easily move the LLC from
Arizona to Ohio. Therefore, I must disolve the LLC in Arizona and create
it in Ohio when I am ready. For now, this means that the site will need
to be a "coming soon" site because the business no longer exists.